snowbridge_pallet_system_frontend::pallet

Trait Config

pub trait Config: Config + Config<RuntimeEvent: From<Event<Self>>> {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RegisterTokenOrigin: EnsureOriginWithArg<Self::RuntimeOrigin, Location, Success = Location>;
    type XcmSender: SendXcm;
    type AssetTransactor: TransactAsset;
    type XcmExecutor: ExecuteXcm<Self::RuntimeCall> + FeeManager;
    type EthereumLocation: Get<Location>;
    type Swap: Swap<Self::AccountId, AssetKind = Location, Balance = u128>;
    type BridgeHubLocation: Get<Location>;
    type UniversalLocation: Get<InteriorLocation>;
    type PalletLocation: Get<InteriorLocation>;
    type AccountIdConverter: ConvertLocation<Self::AccountId>;
    type BackendWeightInfo: BackendWeightInfo;
    type WeightInfo: WeightInfo;
    type Helper: BenchmarkHelper<Self::RuntimeOrigin, Self::AccountId>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type RegisterTokenOrigin: EnsureOriginWithArg<Self::RuntimeOrigin, Location, Success = Location>

Origin check for XCM locations that can register token

type XcmSender: SendXcm

XCM message sender

type AssetTransactor: TransactAsset

To withdraw and deposit an asset.

type XcmExecutor: ExecuteXcm<Self::RuntimeCall> + FeeManager

To charge XCM delivery fees

type EthereumLocation: Get<Location>

Fee asset for the execution cost on ethereum

type Swap: Swap<Self::AccountId, AssetKind = Location, Balance = u128>

To swap the provided tip asset for

type BridgeHubLocation: Get<Location>

Location of bridge hub

type UniversalLocation: Get<InteriorLocation>

Universal location of this runtime.

type PalletLocation: Get<InteriorLocation>

InteriorLocation of this pallet.

type AccountIdConverter: ConvertLocation<Self::AccountId>

type BackendWeightInfo: BackendWeightInfo

Weights for dispatching XCM to backend implementation of register_token

type WeightInfo: WeightInfo

Weights for pallet dispatchables

type Helper: BenchmarkHelper<Self::RuntimeOrigin, Self::AccountId>

A set of helper functions for benchmarking.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.